flaskProject/requirements.txt
Flask~=1.1.2
requests~=2.25.1
kafka-python
flask_cors
choicesenum~=0.7.0